comment-logger/web/requirements.txt
Flask
flask_restful
pymongo
flask_bcrypt